# docs/swagger.yaml
basePath: /financial_tracker
definitions:
  api.ErrorResponse:
    properties:
      error: {}
    type: object
  api.SuccessResponse:
    properties:
      value: {}
    type: object
...
  transactionHandlers.RequestCreateTransaction:
    properties:
      category_id:
        example: 2
        type: integer
      description:
        example: spending on food
        type: string
      limit:
        example: 1000
        type: integer
      name:
        example: jonn
        type: string
    required:
    - category_id
    - limit
    - name
    type: object
  transactionHandlers.RequestUpdateTransaction:
    properties:
      description:
        example: going to a restaurant
        type: string
      limit:
        example: 2000
        type: integer
      name:
        example: jonn
        type: string
      transaction_id:
        example: 1
        type: integer
    required:
    - limit
    - name
    - transaction_id
    type: object
  userHandlers.RefreshToken:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  userHandlers.UserRegistration:
    properties:
      email:
        example: jonn@gmail.com
        type: string
      name:
        example: jonn
        type: string
      password:
        example: securitycod123
        type: string
    required:
    - email
    - name
    - password
    type: object
  userHandlers.UserRequest:
    properties:
      email:
        example: jonn@gmail.com
        type: string
      password:
        example: securitycod123
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8080
info:
  contact:
    email: asssv0423348@gmail.com
    name: Ярослав
    url: https://github.com/goggle-source
  description: API для работы с финансовым трекером
  license:
    name: MIT
  title: Финансовый Трекер
  version: "1.0"
paths:
...
  /registration/get_access_token:
    post:
      consumes:
      - application/json
      description: Получение токена
      parameters:
      - description: для получение access токена
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/userHandlers.RefreshToken'
      produces:
      - application/json
      responses:
        "200":
          description: Получение access токена
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректные входные данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Получение токена
      tags:
      - registration
  /registration/login:
    post:
      consumes:
      - application/json
      description: Вход пользователя в систему
      parameters:
      - description: Данные для авторизации пользователя
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/userHandlers.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Авторизация пользователя
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - registration
  /registration/register:
    post:
      consumes:
      - application/json
      description: Создание нового пользователя
      parameters:
      - description: Данные для регистрации пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/userHandlers.UserRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: Регистрация пользователя
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - registration
  /transaction/create:
    post:
      consumes:
      - application/json
      description: Создание 1 транзакции для пользователя
      parameters:
      - description: данные для создание пользователя
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/transactionHandlers.RequestCreateTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: Транзакция создана успешно
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Ошибка авторизации
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - jwtAuth: []
      summary: Создание транзакции
      tags:
      - transaction
  /transaction/delete:
    delete:
      description: Удаление 1 транзакции для 1 пользователя
      parameters:
      - description: id транзакции
        in: path
        name: req
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Транзакция удалена
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Ошибка авторизации
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Транзакция не найдена
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - jwtAuth: []
      summary: Удаление транзакции
      tags:
      - transaction
  /transaction/get:
    get:
      description: Получение 1 транзакции для 1 пользователя
      parameters:
      - description: id транзакции
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: good
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Ошибка авторизации
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Транзакция не найдена
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - jwtAuth: []
      summary: Получение транзакции
      tags:
      - transaction
  /transaction/update:
    put:
      consumes:
      - application/json
      description: Обновление 1 транзакции для 1 пользователя
      parameters:
      - description: Данные для обновление пользователя
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/transactionHandlers.RequestUpdateTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: Транзакция обновлена
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Ошибка авторизации
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Транзакция не найдена
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - jwtAuth: []
      summary: Обновление транзакции
      tags:
      - transaction
  /user/delete:
    post:
      consumes:
      - application/json
      description: Удаление пользователя
      parameters:
      - description: данные для удаление пользователя
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/userHandlers.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Удаление пользователя
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - jwtAuth: []
      summary: Удаление пользователя
      tags:
      - User
securityDefinitions:
  jwtAuth:
    description: 'type "Bearer" после пробел и jwt token, пример: "Bearer zpdgjeawzgp0398tuP29R0J20THVTP9235BHRNr312r346as2..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"